+++ homepage/design-system/src/app/components/molecules/ThemeToggle.tsx
@@ -19,7 +19,7 @@ export function ThemeToggle({ className }: { className?: string }) {
       type="button"
       className={clsx(
         className,
-        "md:p-2 md:rounded-full md:border",
+        "md:p-1.5 md:rounded-full md:border",
         "text-stone-400 hover:text-stone-900 dark:text-stone-400 dark:hover:text-white",
         "md:hover:bg-stone-200 md:dark:hover:bg-stone-900",
         "transition-colors",
@@ -29,13 +29,13 @@ export function ThemeToggle({ className }: { className?: string }) {
     >
       <MoonIcon
         size={24}
-        strokeWidth={1.5}
-        className="size-5 md:size-6 stroke-stone-900 dark:hidden"
+        strokeWidth={2}
+        className="size-5 stroke-stone-900 dark:hidden"
       />
       <SunIcon
         size={24}
-        strokeWidth={1.5}
-        className="size-5 md:size-6 hidden stroke-white dark:block"
+        strokeWidth={2}
+        className="size-5 hidden stroke-white dark:block"
       />
     </button>
   );