+++ homepage/homepage/lib/docNavigationItems.ts
@@ -12,36 +12,33 @@ export const docNavigationItems = [
       {
         name: "Guide",
         href: "/docs/guide",
-        done: 50,
+        done: {
+          react: 100,
+          "react-native": 0,
+          vue: 0,
+          svelte: 0,
+        },
       },
     ],
   },
   {
-    name: "Project Setup",
+    name: "Project setup",
     items: [
+      {
+        name: "Installation",
+        href: "/docs/project-setup",
+        done: {
+          react: 100,
+          vue: 100,
+          "react-native": 100,
+          svelte: 0,
+        },
+      },
       {
         // jazz mesh, setting api key, free plan, unlimited
         name: "Sync and storage",
         href: "/docs/sync-and-storage",
         done: 100,
-      },
-      {
-        name: "React",
-        href: "/docs/project-setup/react",
-        done: 80,
-      },
-      {
-        name: "React Native",
-        href: "/docs/project-setup/react-native",
-      },
-      {
-        name: "Next.js",
-        href: "/docs/project-setup/next",
-      },
-      {
-        name: "VueJS",
-        href: "/docs/project-setup/vue",
-        done: 80,
       },
       {
         name: "Node.JS / server workers",
@@ -111,12 +108,47 @@ export const docNavigationItems = [
     ],
   },
   {
-    name: "Authentication",
+    name: "Authentication methods",
     items: [
       {
-        name: "Auth methods overview",
+        name: "Overview",
         href: "/docs/authentication/auth-methods",
-        done: 0,
+        done: {
+          react: 50,
+          vue: 0,
+          "react-native": 0,
+          svelte: 0,
+        },
+      },
+      {
+        name: "Passphrase",
+        href: "/docs/authentication/auth-methods#passphrase",
+        done: {
+          react: 50,
+          vue: 0,
+          "react-native": 0,
+          svelte: 0,
+        },
+      },
+      {
+        name: "Passkey",
+        href: "/docs/authentication/auth-methods#passkey",
+        done: {
+          react: 50,
+          vue: 0,
+          "react-native": 0,
+          svelte: 0,
+        },
+      },
+      {
+        name: "Clerk",
+        href: "/docs/authentication/auth-methods#clerk",
+        done: {
+          react: 50,
+          vue: 0,
+          "react-native": 0,
+          svelte: 0,
+        },
       },
       {
         name: "Writing your own",