+++ homepage/homepage/mdx-components.tsx
@@ -1,14 +1,9 @@
+import { DocsLink } from "@/components/docs/DocsLink";
 import type { MDXComponents } from "mdx/types";
-import Link from "next/link";
 
 export function useMDXComponents(components: MDXComponents): MDXComponents {
   return {
-    a: (props) =>
-      props.href ? (
-        <Link href={props.href}>{props.children}</Link>
-      ) : (
-        props.children
-      ),
+    a: (props) => <DocsLink {...props} />,
     ...components,
   };
 }