+++ packages/jazz-browser/src/auth/PasskeyAuth.ts
@@ -13,6 +13,17 @@ type LocalStorageData = {
 
 const localStorageKey = "jazz-logged-in-secret";
 
+/**
+ * `BrowserPasskeyAuth` provides a `JazzAuth` object for passkey authentication.
+ *
+ * ```ts
+ * import { BrowserPasskeyAuth } from "jazz-browser";
+ *
+ * const auth = new BrowserPasskeyAuth(driver, appName);
+ * ```
+ *
+ * @category Auth Providers
+ */
 export class BrowserPasskeyAuth implements AuthMethod {
   constructor(
     public driver: BrowserPasskeyAuth.Driver,
@@ -29,6 +40,9 @@ export class BrowserPasskeyAuth implements AuthMethod {
     this.driver.onError(error);
   }
 
+  /**
+   * @returns A `JazzAuth` object
+   */
   async start(crypto: CryptoProvider): Promise<AuthResult> {
     if (localStorage[localStorageKey]) {
       const localStorageData = JSON.parse(
@@ -90,6 +104,8 @@ export class BrowserPasskeyAuth implements AuthMethod {
                     pubKeyCredParams: [{ alg: -7, type: "public-key" }],
                     authenticatorSelection: {
                       authenticatorAttachment: "platform",
+                      requireResidentKey: true,
+                      residentKey: "required",
                     },
                     timeout: 60000,
                     attestation: "direct",
@@ -171,7 +187,7 @@ export class BrowserPasskeyAuth implements AuthMethod {
   }
 }
 
-/** @category Auth Providers */
+/** @internal */
 // eslint-disable-next-line @typescript-eslint/no-namespace
 export namespace BrowserPasskeyAuth {
   export interface Driver {