+++ packages/jazz-svelte/src/lib/index.ts
@@ -1,0 +1,4 @@
+export { createJazzApp } from "./jazz.svelte.js";
+
+export { createInviteLink, parseInviteLink } from "jazz-browser";
+export * from "./auth/index.js";