+++ packages/jazz-tools/src/tests/subscribe.test.ts
@@ -1,23 +1,7 @@
-const Crypto = await WasmCrypto.create();
-import { connectedPeers } from "cojson/src/streamUtils.js";
 import { describe, expect, it, onTestFinished, vi } from "vitest";
-import {
-  Account,
-  CoFeed,
-  CoList,
-  CoMap,
-  WasmCrypto,
-  co,
-  createJazzContext,
-  fixedCredentialsAuth,
-  isControlledAccount,
-} from "../index.web.js";
-import {
-  FileStream,
-  Group,
-  randomSessionProvider,
-  subscribeToCoValue,
-} from "../internal.js";
+import { Account, CoFeed, CoList, CoMap, co } from "../index.web.js";
+import { FileStream, Group, subscribeToCoValue } from "../internal.js";
+import { setupAccount, waitFor } from "./utils.js";
 
 class ChatRoom extends CoMap {
   messages = co.ref(MessagesList);
@@ -32,34 +16,6 @@ class Message extends CoMap {
 
 class MessagesList extends CoList.Of(co.ref(Message)) {}
 class ReactionsStream extends CoFeed.Of(co.string) {}
-
-async function setupAccount() {
-  const me = await Account.create({
-    creationProps: { name: "Hermes Puggington" },
-    crypto: Crypto,
-  });
-
-  const [initialAsPeer, secondPeer] = connectedPeers("initial", "second", {
-    peer1role: "server",
-    peer2role: "client",
-  });
-
-  if (!isControlledAccount(me)) {
-    throw "me is not a controlled account";
-  }
-  me._raw.core.node.syncManager.addPeer(secondPeer);
-  const { account: meOnSecondPeer } = await createJazzContext({
-    auth: fixedCredentialsAuth({
-      accountID: me.id,
-      secret: me._raw.agentSecret,
-    }),
-    sessionProvider: randomSessionProvider,
-    peersToLoadFrom: [initialAsPeer],
-    crypto: Crypto,
-  });
-
-  return { me, meOnSecondPeer };
-}
 
 function createChatRoom(me: Account | Group, name: string) {
   return ChatRoom.create(
@@ -331,31 +287,3 @@ describe("subscribeToCoValue", () => {
     expect(lastValue.messages[1]).toBe(initialValue.messages[1]);
   });
 });
-
-function waitFor(callback: () => boolean | void) {
-  return new Promise<void>((resolve, reject) => {
-    const checkPassed = () => {
-      try {
-        return { ok: callback(), error: null };
-      } catch (error) {
-        return { ok: false, error };
-      }
-    };
-
-    let retries = 0;
-
-    const interval = setInterval(() => {
-      const { ok, error } = checkPassed();
-
-      if (ok !== false) {
-        clearInterval(interval);
-        resolve();
-      }
-
-      if (++retries > 10) {
-        clearInterval(interval);
-        reject(error);
-      }
-    }, 100);
-  });
-}