+++ packages/cojson-storage-indexeddb/CHANGELOG.md
@@ -1,5 +1,19 @@
 # cojson-storage-indexeddb
 
+## 0.8.34
+
+### Patch Changes
+
+- Updated dependencies [e4f110f]
+  - cojson@0.8.34
+
+## 0.8.32
+
+### Patch Changes
+
+- Updated dependencies [df42b2b]
+  - cojson@0.8.32
+
 ## 0.8.31
 
 ### Patch Changes