+++ packages/jazz-browser-auth-clerk/CHANGELOG.md
@@ -1,5 +1,32 @@
 # jazz-browser-media-images
 
+## 0.8.34
+
+### Patch Changes
+
+- Updated dependencies [e4f110f]
+  - cojson@0.8.34
+  - jazz-browser@0.8.34
+  - jazz-tools@0.8.34
+
+## 0.8.33
+
+### Patch Changes
+
+- Updated dependencies [3cb27e1]
+  - jazz-browser@0.8.33
+
+## 0.8.32
+
+### Patch Changes
+
+- 1a4bda0: Document usage in readme
+- Updated dependencies [df42b2b]
+- Updated dependencies [df42b2b]
+  - cojson@0.8.32
+  - jazz-tools@0.8.32
+  - jazz-browser@0.8.32
+
 ## 0.8.31
 
 ### Patch Changes