+++ packages/jazz-browser/CHANGELOG.md
@@ -1,5 +1,32 @@
 # jazz-browser
 
+## 0.8.34
+
+### Patch Changes
+
+- Updated dependencies [e4f110f]
+  - cojson@0.8.34
+  - cojson-storage-indexeddb@0.8.34
+  - cojson-transport-ws@0.8.34
+  - jazz-tools@0.8.34
+
+## 0.8.33
+
+### Patch Changes
+
+- 3cb27e1: Set created passkey credentials as discoverable
+
+## 0.8.32
+
+### Patch Changes
+
+- Updated dependencies [df42b2b]
+- Updated dependencies [df42b2b]
+  - cojson@0.8.32
+  - jazz-tools@0.8.32
+  - cojson-storage-indexeddb@0.8.32
+  - cojson-transport-ws@0.8.32
+
 ## 0.8.31
 
 ### Patch Changes