+++ packages/jazz-react-auth-clerk/CHANGELOG.md
@@ -1,5 +1,36 @@
 # jazz-browser-media-images
 
+## 0.8.34
+
+### Patch Changes
+
+- 9ca25d1: Rename passkey and clerk examples
+- Updated dependencies [e4f110f]
+  - cojson@0.8.34
+  - jazz-browser-auth-clerk@0.8.34
+  - jazz-react@0.8.34
+  - jazz-tools@0.8.34
+
+## 0.8.33
+
+### Patch Changes
+
+- jazz-browser-auth-clerk@0.8.33
+- jazz-react@0.8.33
+
+## 0.8.32
+
+### Patch Changes
+
+- 1a4bda0: Document usage in readme
+- Updated dependencies [df42b2b]
+- Updated dependencies [1a4bda0]
+- Updated dependencies [df42b2b]
+  - cojson@0.8.32
+  - jazz-browser-auth-clerk@0.8.32
+  - jazz-tools@0.8.32
+  - jazz-react@0.8.32
+
 ## 0.8.31
 
 ### Patch Changes