+++ packages/jazz-run/CHANGELOG.md
@@ -1,5 +1,27 @@
 # jazz-run
 
+## 0.8.34
+
+### Patch Changes
+
+- Updated dependencies [e4f110f]
+  - cojson@0.8.34
+  - cojson-storage-sqlite@0.8.34
+  - cojson-transport-ws@0.8.34
+  - jazz-tools@0.8.34
+
+## 0.8.32
+
+### Patch Changes
+
+- 2e2a13f: Add --json option to output the account data as json
+- Updated dependencies [df42b2b]
+- Updated dependencies [df42b2b]
+  - cojson@0.8.32
+  - jazz-tools@0.8.32
+  - cojson-storage-sqlite@0.8.32
+  - cojson-transport-ws@0.8.32
+
 ## 0.8.31
 
 ### Patch Changes